Create the uc-dissertation class file.
This is the bare bones version of a class file I
am creating which, when complete, will implement
all the specifications of the University of
Cincinnati's graduate school regarding Ph.D.
dissertations.

Right now, all this class file does is load the
KOMA-Script book class (scrbook) with some
specific options.