Add complex titlepage macros involving looping
In order to properly typeset the titlepage, we
need to take the user input for
-previousdegrees
-committee
and reformat it and store this in their starred
variants. The problem is that in both cases, we
don't know *how many* the user will have (degrees
or committee members).

In order to circumvent this problem, we use the
xparse package so that we can allow the user to
input this information in semi-colon separated
lists on which xparse then performs its magic.
Unfortunately, xparse doesn't give us any very
nice ways to access the subarguments, and hence
we have to proceed with a very ugly looping
procedure. It's terrible code, but LaTeX sucks.