Add option for fancy chapter headings, etc.
These changes provide an option for using premade
chapter headings from the fncychap package. The
reason I put these into the class instead of
letting the user do them on their own is that I
wanted to make the fonts match, which is a little
more work (not much) than just loading the package
with the desired option.

However, this is also why I made it only a package
option and it is not forced. That way the user can
modify it as he chooses if they are so inclined.