Add the template for users to get a quick start
This template incorporates all the major features
of the class file. If users can compile this, they
should be able to start writing their dissertation
immediately.

Soon this will be supplemented by the some
documentation which describes all the features
implemented by the class file and the way that
the template is structured.