Fix page numbering problem due to \mainmatter
Since we have to use the mainmatter command twice
in the dissertation file, it is vital that this
doesn't reset the page counter the second time. To
circumvent this we create a command \mainmattertwo
which is defined exactly like \mainmatter except
that it doesn't reset the page counter. Also, with
this setup spacing can be correctly achieved by
using the 'comply' package option.