Improvements to date handling, titlepage spacing
Now using the 'datetime' package in order to
easily typeset just the month and year, with the
month typeset in English.

Remove the optional spacing argument for the
tabular environment on the titlepage. Also, ensure
that the use of the UC logo meets the branding
standards outlined at:
http://www.uc.edu/content/dam/uc/ucomm/docs/UCBrandingStandards.pdf
In particular we ensured there was "safe space"
around the logo.